requirements.txt
gspread 
google-auth 
requests 
python-dotenv
uvicorn fastapi
selenium